+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,8 +1,9 @@
 # 5.1.0 (unreleased)
 
--   Fixed default value for `<Table />` component `onQueryChange` prop.Fixed default value for `<Table />` component `onQueryChange` prop.
+-   Fixed default value for `<Table />` component `onQueryChange` prop.
 -   Deprecate our bespoke component `useFilters` in favor of using the WordPress variety `withFilters`.
 -   Fix screen reader text in `<AdvancedFilters />`.
+-   Added `<AttributeFilter />` component to `<AdvancedFilters />`.
 
 # 5.0.0